moved app.py to front end folder
